refactor(server): refactor utils
